requirements.txt
django-heroku
dj_database_url
gunicorn
psycopg2
djangorestframework
django-ckeditor
django-star-ratings
pillow
whitenoise
python-magic
django-utils-six
djangorestframework_simplejwt
dj-rest-auth 
django-allauth 
django-cors-headers
